+++ examples/stencil-components/vue3-vite/src/App.vue
@@ -7,6 +7,7 @@ import Dropdown from './components/Dropdown.vue'
   <Button />
   <Checkbox />
   <ProgressBar />
+  <TextInput />
 </template>
 
 
@@ -15,7 +16,11 @@ import Alert from './components/Alert.vue'
 import Button from './components/Button.vue'
 import Checkbox from './components/Checkbox.vue'
 import ProgressBar from './components/ProgressBar.vue'
+
 import Dropdown from './components/Dropdown.vue'
+
+import TextInput from './components/TextInput.vue'
+
 
 </script>
 <style scoped>