+++ examples/wrapper-components/react-javascript/src/App.js
@@ -1,44 +1,63 @@
 import './App.css';
-import { useState } from 'react'
-import { IfxSearchBar, IfxProgressBar, IfxButton } from '@infineon/infineon-design-system-react';
+import Link from './components/Link/Link';
+import Button from './components/Button/button';
+import TextField from './components/TextFIeld/TextField';
+import ProgressBar from './components/ProgressBar/ProgressBar';
+import SearchBar from './components/SearchBar/SearchBar';
+import Accordion from './components/Accordion/Accordion';
+import RadioButton from './components/RadioButton/RadioButton';
 
+import Checkbox from './components/Checkbox/Checkbox';
+import Spinner from './components/Spinner/Spinner';
+import NumberIndicator from './components/NumberIndicator/NumberIndicator';
 
 function App() {
-  const [progressValue, setProgressValue] = useState(25);  // Add this line
-
-  // Define your methods here
-  const updateProgressOnClick = () => {
-    setProgressValue((currentValue) => {
-      console.log("updating progress-bar value ", progressValue)
-      return currentValue < 100 ? currentValue += 10 : currentValue = 10;
-    });
-  };
-  const handleSearch = (event) => {
-    console.log("handling search: ", event.detail?.detail)
-  };
-
-
-  const handleMouseOver = () => {
-    console.log('Mouse over event detected');
-  };
-
   return (
     <div>
-      <h1 className="header">Stencil Framework integration - React + JS</h1>
-      <h2>Ifx-Search-Bar</h2>
-      <IfxSearchBar onIfxChange={handleSearch} style={{ width: '100%' }} show-close-button="true"></IfxSearchBar>
+      <h1>Stencil Framework integration - React + JS</h1>
+
+      <h2>Search Bar</h2>
+      <SearchBar />
       <br />
 
-      <h2>Ifx-Progress-Bar</h2>
-      <IfxProgressBar value={progressValue} size="m" show-label="true"></IfxProgressBar>
+      <h2>Button</h2>
+      <Button />
       <br />
 
-      <IfxButton variant="outline" icon="" position="left" href="" target="_blank" color="primary" size="m"
-        disabled={false} onClick={updateProgressOnClick} onMouseOver={handleMouseOver}>
-        Increase by 10
-      </IfxButton>
+      <h2>Progress Bar</h2>
+      <ProgressBar />
+      <br />
+
+      <h2>Text Field</h2>
+      <TextField />
+      <br />
+
+      <h2>Accordion</h2>
+      <Accordion />
+      <br />
+
+      <h2>Radio Button</h2>
+      <RadioButton />
+      <br />
+
+      <h2>Checkbox</h2>
+      <Checkbox />
+      <br />
+
+      <h2>Spinner</h2>
+      <Spinner />
+      <br />
+
+      <h2>Link</h2>
+      <Link />
+      <br />
+
+
+      <h2>Number Indicator</h2>
+      <NumberIndicator />
+      <br />
     </div>
-  );
-}
 
+  )
+}
 export default App;