+++ examples/wrapper-components/react-typescript/src/App.tsx
@@ -1,50 +1,63 @@
-import { useRef, useState } from 'react'
-import './App.css'
-import { IfxProgressBar, IfxSearchBar, IfxButton } from '@infineon/infineon-design-system-react';
 
+import './App.css'
+import Link from './components/Link/Link'
+import Button from './components/Button/button';
+import TextField from './components/TextFIeld/TextField';
+import ProgressBar from './components/ProgressBar/ProgressBar';
+import SearchBar from './components/SearchBar/SearchBar';
+import Accordion from './components/Accordion/Accordion';
+import RadioButton from './components/RadioButton/RadioButton';
+
+import './App.css'
+import Spinner from './Components/Spinner/Spinner'
+
+import './App.css'
+import NumberIndicator from './components/NumberIndicator/NumberIndicator'
 
 function App() {
-  const [progressValue, setProgressValue] = useState(25);  // Add this line
-  const ifxButtonRef = useRef(null);
-
-  // Define your methods here
-  const updateProgressOnClick = () => {
-    setProgressValue((currentValue) => {
-      return currentValue < 100 ? currentValue += 10 : currentValue = 10;
-    });
-    console.log("ifxButtonRef. ", ifxButtonRef.current)
-   };
-
-  const handleSearch = (event: CustomEvent) => {
-    console.log("handling search: ", event.detail?.detail)
-  };
-  
-
-  const handleMouseOver = () => {
-    console.log('Mouse over event detected');
-  };
+ 
+ 
 
   return (
-    <>
-      <div>
-    <h1 className="header">Stencil Framework integration - React + TS </h1>
-    <h2>Ifx-Search-Bar</h2>
-    <IfxSearchBar onIfxChange={handleSearch} style={{ width: '100%' }}  show-close-button="true"></IfxSearchBar>
-    <br />
-
-      <h2>Ifx-Progress-Bar</h2>
-      <IfxProgressBar value={progressValue} size="m" show-label="true"></IfxProgressBar>
+    <div>
+      <h1>Stencil Framework integration - React + TS </h1>
+      <h2>Search Bar</h2>
+      <SearchBar />
       <br />
- 
-    <IfxButton ref={ifxButtonRef} variant="outline" icon="" position="left" href="" target="_blank" color="primary" size="m"
-      disabled={false} onClick={updateProgressOnClick} onMouseOver={handleMouseOver}>
-      Increase by 10
-    </IfxButton>
 
-      </div>
-    
-    </>
-  )
+      <h2>Button</h2>
+      <Button />
+      <br />
+
+      <h2>Progress Bar</h2>
+      <ProgressBar />
+      <br />
+
+      <h2>Text Field</h2>
+      <TextField />
+      <br />
+
+      <h2>Accordion</h2>
+      <Accordion />
+      <br />
+
+      <h2>Radio Button</h2>
+      <RadioButton />
+      <br />
+      
+      <h2>Spinner</h2>
+      <Spinner />
+      <br />
+      
+      <h2>Link</h2>
+      <Link />
+      <br />
+
+      <h2>Number Indicator</h2>
+      <NumberIndicator />
+      
+    </div>
+   )
 }
 
 export default App