+++ examples/wrapper-components/vue-javascript/src/App.vue
@@ -1,11 +1,16 @@
 
 <template>
-  <h1 class="header">Using Stencil web components in a Vue application</h1>
+  <h1 class="header">Stencil Framework integration - Vue + JS </h1>
   <Alert />
   <Button />
   <Checkbox />
+  <RadioButton />
   <ProgressBar />
   <Search />
+  <TextInput />
+  <Spinner />
+  <Link />
+  <NumberIndicator />
 </template>
 
 
@@ -13,8 +18,13 @@
 import Alert from './components/Alert.vue'
 import Button from './components/Button.vue'
 import Checkbox from './components/Checkbox.vue'
+import NumberIndicator from './components/NumberIndicator.vue'
 import ProgressBar from './components/ProgressBar.vue'
 import Search from './components/Search.vue'
+import RadioButton from './components/RadioButton.vue'
+import TextInput from './components/TextInput.vue'
+import Spinner from './components/Spinner.vue'
+import Link from './components/Link.vue'
 
 </script>