+++ packages/components/src/components/accordion/accordion.stories.tsx
@@ -3,6 +3,8 @@ import { action } from '@storybook/addon-actions';
 export default {
   title: 'Components/Accordion',
   component: 'ifx-accordion',
+  tags: ['autodocs'],
+
   argTypes: {
     amountOfItems: { control: 'number' },
   },
@@ -21,8 +23,8 @@ const Template = (args) => {
         Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent volutpat, ligula eu aliquam bibendum, orci nisl cursus ipsum, nec egestas odio sapien eget neque.
       </p>
     `;
-    item.addEventListener('ifxItemOpened', action('ifxItemOpened'));
-    item.addEventListener('ifxItemClosed', action('ifxItemClosed'));
+    item.addEventListener('ifxItemOpen', action('ifxItemOpen'));
+    item.addEventListener('ifxItemClose', action('ifxItemClose'));
 
     accordionElement.append(item);
   }