+++ packages/components/src/components/checkbox/checkbox.tsx
@@ -7,19 +7,21 @@ import { Component, h, Prop, Element, State, Event, EventEmitter, Watch } from '
 })
 
 export class Checkbox {
+  private inputElement: HTMLInputElement;
+
   @Element() el;
   @Prop() disabled: boolean = false;
   @Prop() value: boolean = false;
   @Prop() error: boolean = false;
   @Prop() name: string = '';
-  @State() hasSlot: boolean = true;
   @State() internalValue: boolean;
-  @Event({ eventName: 'ifxChange' }) ifxChange: EventEmitter;
+  @Event({ bubbles: true, composed: true }) ifxChange: EventEmitter;
 
   handleCheckbox() {
     if (!this.disabled) {
       this.internalValue = !this.internalValue;
-      this.ifxChange.emit({ value: this.internalValue });
+      this.inputElement.checked = this.internalValue; // update the checkbox's checked property
+      this.ifxChange.emit(this.internalValue);
     }
   }
 
@@ -27,15 +29,10 @@ export class Checkbox {
   valueChanged(newValue: boolean, oldValue: boolean) {
     if (newValue !== oldValue) {
       this.internalValue = newValue;
+      this.inputElement.checked = this.internalValue; // update the checkbox's checked property
     }
   }
 
-  componentDidUpdate() {
-    const slot = this.el.innerHTML;
-    if (slot) {
-      this.hasSlot = true;
-    } else this.hasSlot = false;
-  }
 
   handleKeydown(event) {
     // Keycode 32 corresponds to the Space key, 13 corresponds to the Enter key
@@ -46,17 +43,30 @@ export class Checkbox {
   }
 
   componentWillLoad() {
-    this.internalValue = this.value;
+    this.internalValue = this.internalValue || false
   }
 
-
   render() {
+    const slot = this.el.innerHTML;
+    let hasSlot = false;
+
+    if (slot) {
+      hasSlot = true;
+    }
+
     return (
       <div class="checkbox__container">
-        <input type="checkbox" hidden
+        <input
+          type="checkbox"
+          hidden
+          ref={(el) => (this.inputElement = el)}
           name={this.name}
           checked={this.internalValue}
-          value={`${this.internalValue}`} />
+          onChange={this.handleCheckbox.bind(this)} // Listen for changes here
+          id='checkbox'
+          value={`${this.internalValue}`}
+        />
+
         <div
           tabindex="0"
           onClick={this.handleCheckbox.bind(this)}
@@ -71,7 +81,7 @@ export class Checkbox {
         ${this.error ? 'error' : ""}`}>
           {this.internalValue && <ifx-icon icon="check-12"></ifx-icon>}
         </div>
-        {this.hasSlot &&
+        {hasSlot &&
           <div id="label" class={`label ${this.error ? 'error' : ""} ${this.disabled ? 'disabled' : ""} `} onClick={this.handleCheckbox.bind(this)}>
             <slot />
           </div>}