+++ packages/components/src/components/text-input/text-input.tsx
@@ -1,22 +1,38 @@
-import { Component, h, Element, Prop } from '@stencil/core';
+import { Component, h, Event, Element, Prop, EventEmitter, Watch } from '@stencil/core';
 
 @Component({
   tag: 'ifx-text-input',
-  styleUrl: 'text-input.scss', 
+  styleUrl: 'text-input.scss',
   shadow: true
 })
 
 export class TextInput {
+  private inputElement: HTMLInputElement;
   @Element() el;
   @Prop() placeholder: string = "Placeholder"
-
-  @Prop() value: string = ""
+  @Prop({ mutable: true }) value: string = '';
   @Prop() error: boolean = false;
   @Prop() errorMessage: string = ""
   @Prop() success: boolean = false;
   @Prop() disabled: boolean = false;
   @Prop() readonly: boolean = false;
   @Prop() icon: boolean = false;
+  @Event() ifxInput: EventEmitter<String>;
+
+
+  @Watch('value')
+  valueWatcher(newValue: string) {
+    if (newValue !== this.inputElement.value) {
+      this.inputElement.value = newValue;
+    }
+  }
+
+
+  handleInput() {
+    const query = this.inputElement.value;
+    this.value = query; // update the value property when input changes
+    this.ifxInput.emit(this.value);
+  }
 
   render() {
     return (
@@ -27,20 +43,20 @@ export class TextInput {
           </label>
         </div>
         <div class="textInput__bottom-wrapper">
-          <input 
-
-          readonly={this.readonly}
-          disabled={this.disabled}
-          type="text" 
-          id='text-field' 
-          value={this.value}
-
-          placeholder={this.placeholder} 
-          class={`${this.error ? 'error' : ""} ${this.success ? "success" : ""}`}/>
-          {this.error && 
+          <input
+            ref={(el) => (this.inputElement = el)}
+            readonly={this.readonly}
+            disabled={this.disabled}
+            type="text"
+            id='text-field'
+            value={this.value}
+            onInput={() => this.handleInput()}
+            placeholder={this.placeholder}
+            class={`${this.error ? 'error' : ""} ${this.success ? "success" : ""}`} />
+          {this.error &&
             <div class="textInput__bottom-wrapper-error">
-            {this.errorMessage}
-          </div>}
+              {this.errorMessage}
+            </div>}
 
           {this.icon && <ifx-icon icon='chevron-down-16'></ifx-icon>}