+++ packages/components/CHANGELOG.md
@@ -1,3 +1,200 @@
+# v18.2.15 (Mon Jul 10 2023)
+
+#### 🐛 Bug Fix
+
+- 218-add-about-and-faq-section-to-storybook [#221](https://github.com/Infineon/infineon-design-system-stencil/pull/221) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.14 (Mon Jul 10 2023)
+
+#### 🐛 Bug Fix
+
+- 185 review number indicator [#237](https://github.com/Infineon/infineon-design-system-stencil/pull/237) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+- 209 review link [#240](https://github.com/Infineon/infineon-design-system-stencil/pull/240) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### 🔩 Dependency Updates
+
+- Bump tough-cookie from 4.1.2 to 4.1.3 in /examples/wrapper-components/react-javascript [#242](https://github.com/Infineon/infineon-design-system-stencil/pull/242) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump tough-cookie from 4.1.2 to 4.1.3 in /examples/stencil-components/react [#243](https://github.com/Infineon/infineon-design-system-stencil/pull/243) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+#### Authors: 3
+
+- [@dependabot[bot]](https://github.com/dependabot[bot])
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.13 (Mon Jul 10 2023)
+
+#### 🐛 Bug Fix
+
+- 191 review spinner [#214](https://github.com/Infineon/infineon-design-system-stencil/pull/214) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.12 (Fri Jul 07 2023)
+
+#### 🐛 Bug Fix
+
+- 187 review radio button [#210](https://github.com/Infineon/infineon-design-system-stencil/pull/210) (lechneve@ISCN5CG1201S7M.infineon.com [@SergejLust](https://github.com/SergejLust) [@tishoyanchev](https://github.com/tishoyanchev) [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 4
+
+- [@SergejLust](https://github.com/SergejLust)
+- [@verena-ifx](https://github.com/verena-ifx)
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.11 (Tue Jul 04 2023)
+
+#### 🐛 Bug Fix
+
+- 206-research-the-possibility-to-include-assets-in-mdx-documentation-files [#217](https://github.com/Infineon/infineon-design-system-stencil/pull/217) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.10 (Fri Jun 30 2023)
+
+#### 🐛 Bug Fix
+
+- 205 review button add form support [#207](https://github.com/Infineon/infineon-design-system-stencil/pull/207) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.9 (Fri Jun 30 2023)
+
+#### 🐛 Bug Fix
+
+- 212 missing icon preview stories file [#213](https://github.com/Infineon/infineon-design-system-stencil/pull/213) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.8 (Thu Jun 22 2023)
+
+#### 🐛 Bug Fix
+
+- 193 review textinput [#200](https://github.com/Infineon/infineon-design-system-stencil/pull/200) (lechneve@ISCN5CG1201S7M.infineon.com [@tishoyanchev](https://github.com/tishoyanchev) [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 3
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.7 (Wed Jun 21 2023)
+
+#### 🐛 Bug Fix
+
+- Button fixes [#199](https://github.com/Infineon/infineon-design-system-stencil/pull/199) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.6 (Tue Jun 20 2023)
+
+#### 🐛 Bug Fix
+
+- 186 review progress bar [#203](https://github.com/Infineon/infineon-design-system-stencil/pull/203) (lechneve@ISCN5CG1201S7M.infineon.com [@tishoyanchev](https://github.com/tishoyanchev) [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 3
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.5 (Mon Jun 19 2023)
+
+#### 🐛 Bug Fix
+
+- 160 e2e tests [#197](https://github.com/Infineon/infineon-design-system-stencil/pull/197) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.2.4 (Fri Jun 16 2023)
+
+#### 🐛 Bug Fix
+
+- Accordion issues fixed [#178](https://github.com/Infineon/infineon-design-system-stencil/pull/178) ([@tishoyanchev](https://github.com/tishoyanchev))
+
+#### Authors: 1
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+
+---
+
+# v18.2.3 (Fri Jun 16 2023)
+
+#### 🐛 Bug Fix
+
+- 159 restructuring breadcrumb [#172](https://github.com/Infineon/infineon-design-system-stencil/pull/172) ([@tishoyanchev](https://github.com/tishoyanchev))
+
+#### Authors: 1
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+
+---
+
+# v18.2.2 (Wed Jun 14 2023)
+
+#### 🐛 Bug Fix
+
+- Alert close event [#194](https://github.com/Infineon/infineon-design-system-stencil/pull/194) ([@tishoyanchev](https://github.com/tishoyanchev))
+
+#### 🔩 Dependency Updates
+
+- Bump vite from 4.3.8 to 4.3.9 in /examples/stencil-components/vanilla-cdn [#169](https://github.com/Infineon/infineon-design-system-stencil/pull/169) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump vite from 4.3.8 to 4.3.9 in /examples/wrapper-components/react-typescript [#168](https://github.com/Infineon/infineon-design-system-stencil/pull/168) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump vite from 4.3.8 to 4.3.9 in /examples/wrapper-components/vue-javascript [#167](https://github.com/Infineon/infineon-design-system-stencil/pull/167) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+#### Authors: 2
+
+- [@dependabot[bot]](https://github.com/dependabot[bot])
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+
+---
+
 # v18.2.1 (Tue Jun 13 2023)
 
 #### 🐛 Bug Fix